@@ -23,8 +23,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 }
 
@@ -37,4 +37,11 @@ dependencies {
     testImplementation libs.junit
     androidTestImplementation libs.ext.junit
     androidTestImplementation libs.espresso.core
+    implementation libs.retrofit
+    implementation libs.converter.gson
+    implementation libs.compiler
+    implementation libs.jackson.databind
+    implementation libs.jackson.core
+    implementation libs.okhttp
+    implementation libs.logging.interceptor
 }
